Reverted to 1st Jul commit as this implementation was not proper.
Now the contructor is just a way to create objets of the class which are manipulated by various methods. The ListLoader() method reads from the existing file and creates a new object of the class and adds it to the existing LinkedList. All methods have been refactored to work with the type TerminalToDo which results in the class objects being manipulated.

Next to work on is to move all the methods which manilupate the list objects to another class to improve readability of code. Also need to change the comments as a lot of the funcionality has been intricately changed.